finishes first still buggy version of new table grid
